requirements.txt
torch
torchvision
numpy
opencv-python-headless
streamlit
Pillow
matplotlib
gdown